feat(build.sh): add build script to compile Pug templates, Sass stylesheets, and copy JavaScript files to the docs folder
The build.sh script is added to automate the build process. It compiles Pug templates from the src folder to the docs folder using the pug command with the -P flag to preserve whitespace. It also compiles Sass stylesheets from src/style.sass to docs/style.css using the sass command. Additionally, it copies JavaScript files from the src folder to the docs folder using the cp command. The script exits with a status code of 0.